update admin dashboard init

+++ database/migrations/2024_09_09_113700_create_user_attributes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('attribute_name');
-            $table->integer('attribute_value');
+            $table->float('attribute_value');
             $table->timestamps();
         });
     }
